global refactor styles
