--- src/components/ui/UniversalTable/SingleRow/uniSingleRowDeck.module.scss
@@ -1,135 +1,0 @@
-.root {
-  margin-bottom: 24px;
-}
-
-.arrow {
-  width: 24px;
-  height: 24px;
-  transition: transform 0.5s;
-}
-
-.rotate {
-  transform: rotate(180deg);
-}
-
-.empty {
-  padding: 60px 0;
-  text-align: center;
-}
-
-.iconBtns {
-  position: relative;
-
-  display: flex;
-  flex-flow: row nowrap;
-  gap: 10px;
-  justify-content: center;
-
-  .btn {
-    position: relative;
-
-    display: inherit;
-
-    width: 16px;
-    height: 16px;
-    padding: 0;
-
-    background-color: transparent;
-    border: none;
-    outline: none;
-    box-shadow: none;
-
-    &::before {
-      content: '';
-
-      position: absolute;
-      z-index: 0;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%) scale(0);
-
-      width: 28px;
-      height: 28px;
-
-      opacity: 0.5;
-      background-color: transparent;
-      border-radius: 50%;
-
-      transition: 0.5s;
-    }
-
-    &:hover::before,
-    &:active::before,
-    &:focus-visible::before {
-      transform: translate(-50%, -50%) scale(1);
-      background-color: var(--color-dark-100);
-    }
-
-    &:active::before {
-      background-color: var(--color-dark-300);
-    }
-
-    &:disabled::before {
-      background-color: transparent;
-    }
-  }
-}
-
-.playCircleOutline,
-.Edit2Outline,
-.TrashOutline {
-  cursor: pointer;
-
-  position: absolute;
-
-  width: 16px;
-  height: 16px;
-
-  color: var(--color-light-100);
-
-  &:disabled {
-    background-color: transparent;
-  }
-}
-
-.disabled {
-  cursor: default;
-  opacity: 0.5;
-}
-
-// Region
-.nameBlock {
-  display: flex;
-  flex-flow: row nowrap;
-  gap: 10px;
-  justify-content: flex-start;
-
-  height: 62px;
-  padding: 0;
-
-  text-decoration: none;
-  white-space: nowrap;
-
-  background-color: transparent;
-  box-shadow: none;
-
-  &:hover,
-  &:focus {
-    background-color: transparent;
-  }
-
-  .defaltDeckImg {
-    width: 100%;
-    max-width: 70px;
-    height: 100%;
-  }
-}
-
-.container:hover {
-  background-color: var(--color-dark-700);
-}
-
-.authorCell {
-  width: 200px;
-  max-width: 200px;
-}
