--- src/pagesMinin/ModalsForTable/modalOnAddDeckMinin.module.scss
@@ -1,75 +1,0 @@
-//* p {
-//  margin: 0;
-//}
-.body {
-  margin-bottom: 45px;
-  position: relative;
-  z-index: 1;
-}
-.preview {
-  margin-top: 8px;
-  margin-bottom: 16px;
-}
-.uploadImg {
-  position: relative;
-  z-index: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--color-dark-300);
-  border-radius: 4px;
-  padding: 10px 0 10px 0;
-  transition: background-color 0.2s ease-in-out;
-  cursor: pointer;
-  margin-bottom: 45px;
-  box-shadow: none;
-  &.removeImg {
-    margin-bottom: 1rem;
-  }
-  &:hover {
-    background-color: var(--color-dark-100);
-  }
-  &:active {
-    background-color: var(--color-light-900);
-  }
-  &:focus,
-  &:first-child:focus {
-    outline: unset;
-    box-shadow: 0 0 0 2px inset var(--color-info-700);
-  }
-}
-
-.inputImg {
-  //position: absolute;
-  //z-index: -1;
-  //opacity: 0;
-  //width: 100%;
-  opacity: 0;
-  width: 0;
-  height: 0;
-  padding: 0;
-  margin: 0;
-}
-
-.checkbox {
-  margin-left: 5px;
-}
-
-.icon {
-  width: 16px;
-  height: 16px;
-  margin-right: 5px;
-}
-
-.customClass {
-  width: 420px;
-}
-
-.input {
-  width: 100%;
-  margin-bottom: 24px;
-}
-.footer {
-  display: flex;
-  justify-content: space-between;
-}
