+++ src/pagesMinin/ModalsForTable/modalsMinin.module.scss
@@ -51,10 +51,6 @@
   margin: 0;
 }
 
-.checkbox {
-  margin-left: 5px;
-}
-
 .icon {
   width: 20px;
   height: 20px;
