+++ src/pagesMinin/TablesMinin/DeckCardsTableMinin/deckCardsTableMinin.module.scss
@@ -1,0 +1,5 @@
+// Для Cards
+.tableHeadCellCards:nth-child(1),
+.tableHeadCellCards:nth-child(2) {
+  width: 27%;
+}
