+++ src/pagesMinin/TablesMinin/DecksTable/decksTableMinin.module.scss
@@ -1,0 +1,7 @@
+// Для Decks
+.tableHeadCellDecks:nth-child(1) {
+  width: 35%;
+}
+.tableHeadCellDecks:nth-child(2) {
+  width: 10%;
+}
