+++ src/pagesMinin/decksPageMinin.module.scss
@@ -1,30 +1,14 @@
 .common {
   max-width: 1008px;
 }
-// Region ДОБАВИТЬ
 // Общее
 .tableRoot {
   table-layout: fixed;
 }
-// Для Decks
-.tableHeadCellDecks:nth-child(1) {
-  width: 35%;
-}
-.tableHeadCellDecks:nth-child(2) {
-  width: 10%;
-}
 .emptyTableHeadCell {
   width: 15%;
 }
-// Для Cards
-.tableHeadCellCards:nth-child(1),
-.tableHeadCellCards:nth-child(2) {
-  width: 27%;
-}
-//.emptyTableHeadCell {
-//  width: 15%;
-//}
-// End
+
 .heading {
   display: grid;
   gap: 7px;
